pyproject.toml
[project]
name = "ai-powered-workout-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-groq>=0.3.8",
    "langchain-redis>=0.2.4",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "upstash-redis>=1.4.0",
    "uvicorn[standard]>=0.37.0",
]
